frame buffer: memory inferring BRAM now
Before the synthesis tool was not able to infer BRAM as the description was not mappable on it. I've restructured the code, removed the memory reset (which is not necessary).
While it was not really necessary for the 8x6 board, the 32x24 board consumes about 30% of the totally available LUTs which is inacceptable.