readme: v2 works on custom pcb
This is a big success, the results are so good that we can simply develop the full size version without any restrictions. The board is able to clock it much faster than necessary.

The only downside is that we are going to lose brightness. While we could compensate for that by reducing the 56 Ohm resistors to lower values, we want to stay in spec.
Another approach would be to double the row data shift registers effectively duplicating brightness.
This means that instead of 25% brightness, we would end up with 50%.