fullsize led board: fixed backwards wiring
It's not as beautiful as I'd like to have it, but this will suffice, otherwise it will take way too much time to reroute the VIAs.
If I want to improve this design further, I'd like to use SMD resistors anyway, so a full rerout would be necessary.
Lesson learned I guess...