video driver: added emptying of padded bytes
I'm not quite happy with this solution, the video driver became more and more of a tangled mess.
There has to be a better and more straight forward solution for this.

Also deleted the .bak file, which was automatically generated a few commits back from a hexviewer which I missed.