small led board v2: moved buffer pullups to +3.3v of pmod
This was a serious design oversight of not paying attention that the pullups could inject current into the FPGA IO when it's driven high (3.3V).
The only reason the FPGA survived this for so long is that the current was significantly limited by the pullups (10K/1M results in way too weak current) and an internal ESD diode possibly clamped the current.

The fullsize board is also affected and will be fixed in the next commit.